Refactored switch-related code into different modules
