+++ bcipy/signal/model/switch_model.py
