+++ bcipy/simulator/data/processor_registry.py
@@ -7,7 +7,7 @@ from bcipy.signal.model.base_model import SignalModel
 # flake8: noqa
 from bcipy.simulator.data.data_process import (EegRawDataProcessor,
                                                RawDataProcessor)
-from bcipy.simulator.demo.switch_data_processor import SwitchDataProcessor
+from bcipy.simulator.data.switch_data_processor import SwitchDataProcessor
 from bcipy.task.data import EvidenceType
 
 log = logger.getLogger(__name__)
